+++ protocols/bin_state_message.py
@@ -3,8 +3,8 @@ from utils.cords import Cords
 
 
 class BinStateMessage(BaseMessage):
-    fill_level_percentage: int
-    max_volume: int
+    fill_level_percentage: float
+    max_volume: float
     position: Cords
 
     @classmethod
